Beagleboard INTC bugfix, code size & performance: made inline functions static and fixed misuse of volatile keyword in intcRegReadBE.
As a consequence, the inline functions have been removed from the header file. However, they are low-level functions, and should not be used outside the driver anyway.

In intcRegReadBE, a register value was declared non-volatile, and was later read into a normal variable declared volatile. This makes no sense; the keyword is now used on the register value instead.