Make use of security extensions to speed up interrupt handling
Security extensions add a vector base address register (VBAR). The value of this register is an offset, applied to the PC on exception entry. By default the OMAP boot ROM sets it to 0x00014000, causing interrupt handlers to travel from the boot ROM to the SRAM, then from the SRAM to the SDRAM into the hypervisor. Putting the exception vector inside the hypervisor avoids 2 branches and 4 memory accesses for each exception.