andPC is implemented
Resolved 2 bugs:
bug1 -> undefined handler crash.  This was due to the static branch at
the end of the blockCopyCache the first nibble of the instruction was 
set to 0xF instead of 0xE resulting in a branch that changed execution
mode to Thumb

bug2 -> cleanup of BlockCopyCache entry didn't take a splitted block
into account.  This can occure when a block is at the end of the block-
CopyCache.  Now it does (isn't validated yet since the situation hasn't 
occurred yet).